Clean up duplicate codes and make no apps are lost with max
