--- app/src/main/java/com/sduduzog/slimlauncher/adapters/OpenAppAdapter.kt
@@ -1,44 +1,0 @@
-package com.sduduzog.slimlauncher.adapters
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.sduduzog.slimlauncher.R
-import com.sduduzog.slimlauncher.data.model.App
-import com.sduduzog.slimlauncher.utils.OnAppClickedListener
-
-class OpenAppAdapter(private val listener: OnAppClickedListener) : RecyclerView.Adapter<OpenAppAdapter.ViewHolder>() {
-
-    private var apps: List<App> = listOf()
-
-    override fun getItemCount(): Int = apps.size
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val item = apps[position]
-        holder.appName.text = item.appName
-        holder.itemView.setOnClickListener {
-            listener.onAppClicked(item)
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.add_app_fragment_list_item, parent, false)
-        return ViewHolder(view)
-    }
-
-    fun setItems(apps: List<App>){
-        this.apps = apps
-        notifyDataSetChanged()
-    }
-
-    inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-        val appName: TextView = itemView.findViewById(R.id.aa_list_item_app_name)
-
-        override fun toString(): String {
-            return super.toString() + " '${appName.text}'"
-        }
-    }
-}
