+++ app/src/main/java/com/sduduzog/slimlauncher/utils/BaseFragment.kt
@@ -1,15 +1,21 @@
 package com.sduduzog.slimlauncher.utils
 
+import android.content.Context
 import android.content.Intent
+import android.content.pm.LauncherApps
 import android.os.Build
+import android.os.Process
+import android.os.UserManager
 import android.util.TypedValue
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityOptionsCompat
 import androidx.fragment.app.Fragment
+import com.sduduzog.slimlauncher.BuildConfig
 import com.sduduzog.slimlauncher.MainActivity
 import com.sduduzog.slimlauncher.R
+import com.sduduzog.slimlauncher.data.model.App
 
 abstract class BaseFragment : Fragment(), ISubscriber {
 
@@ -63,4 +69,33 @@ abstract class BaseFragment : Fragment(), ISubscriber {
     open fun onBack(): Boolean = false
 
     open fun onHome() {}
+
+    protected fun getInstalledApps(): List<App> {
+        val list = mutableListOf<App>()
+
+        val manager = requireContext().getSystemService(Context.USER_SERVICE) as UserManager
+        val launcher = requireContext().getSystemService(Context.LAUNCHER_APPS_SERVICE) as LauncherApps
+        val myUserHandle = Process.myUserHandle()
+
+        for (profile in manager.userProfiles) {
+            val prefix = if (profile == myUserHandle) "" else "\uD83C\uDD46 " //Unicode for boxed w
+            val profileSerial = manager.getSerialNumberForUser(profile)
+
+            for (activityInfo in launcher.getActivityList(null, profile)) {
+                val app = App(
+                        appName = prefix + activityInfo.label.toString(),
+                        packageName = activityInfo.applicationInfo.packageName,
+                        activityName = activityInfo.name,
+                        userSerial = profileSerial
+                )
+                list.add(app)
+            }
+        }
+
+        list.sortBy{it.appName}
+
+        val filter = mutableListOf<String>()
+        filter.add(BuildConfig.APPLICATION_ID)
+        return list.filterNot { filter.contains(it.packageName) }
+    }
 }
